backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
python-multipart==0.0.9 